mapping refactoring
